+++ module-service/src/main/java/com/checkping/dto/question/comment/TaskBoardCommentRequest.java
@@ -1,4 +1,4 @@
-package com.checkping.dto;
+package com.checkping.dto.question.comment;
 
 import com.checkping.domain.question.TaskBoard;
 import com.checkping.domain.question.TaskBoardComment;
