+++ module-service/src/main/java/com/checkping/service/question/TaskBoardServiceImpl.java
@@ -4,14 +4,14 @@ import com.checkping.domain.question.TaskBoard;
 import com.checkping.domain.question.TaskBoardComment;
 import com.checkping.domain.question.TaskBoardFile;
 import com.checkping.domain.question.TaskBoardLink;
-import com.checkping.dto.TaskBoardLinkRequest;
-import com.checkping.dto.TaskBoardRequest;
-import com.checkping.dto.TaskBoardRequest.RegisterDto;
-import com.checkping.dto.TaskBoardRequest.SearchCondition;
-import com.checkping.dto.TaskBoardRequest.UpdateDto;
-import com.checkping.dto.TaskBoardResponse;
-import com.checkping.dto.TaskBoardResponse.TaskBoardItemDto;
-import com.checkping.dto.TaskBoardResponse.TaskBoardListDto;
+import com.checkping.dto.question.link.TaskBoardLinkRequest;
+import com.checkping.dto.question.TaskBoardRequest;
+import com.checkping.dto.question.TaskBoardRequest.RegisterDto;
+import com.checkping.dto.question.TaskBoardRequest.SearchCondition;
+import com.checkping.dto.question.TaskBoardRequest.UpdateDto;
+import com.checkping.dto.question.TaskBoardResponse;
+import com.checkping.dto.question.TaskBoardResponse.TaskBoardItemDto;
+import com.checkping.dto.question.TaskBoardResponse.TaskBoardListDto;
 import com.checkping.exception.project.TaskBoardNotFoundEntityException;
 import com.checkping.infra.repository.question.TaskBoardReader;
 import com.checkping.infra.repository.question.TaskBoardStore;
