+++ module-service/src/main/java/com/checkping/service/question/link/TaskBoardLinkServiceImpl.java
@@ -2,8 +2,8 @@ package com.checkping.service.question.link;
 
 import com.checkping.domain.question.TaskBoard;
 import com.checkping.domain.question.TaskBoardLink;
-import com.checkping.dto.TaskBoardLinkRequest;
-import com.checkping.dto.TaskBoardLinkResponse.TaskBoardLinkDto;
+import com.checkping.dto.question.link.TaskBoardLinkRequest;
+import com.checkping.dto.question.link.TaskBoardLinkResponse.TaskBoardLinkDto;
 import com.checkping.exception.project.TaskBoardNotFoundEntityException;
 import com.checkping.infra.repository.question.TaskBoardReader;
 import com.checkping.infra.repository.question.link.TaskBoardLinkStore;
